refactor:  main 파일 정리
